update validation stuff
