+++ spec/cc/analyzer/issue/validations/relative_path_validation_spec.rb
@@ -1,6 +1,6 @@
 require "spec_helper"
 
-module CC::Analyzer::Validations
+module CC::Analyzer::Issue::Validations
   describe RelativePathValidation do
     describe "#valid?" do
       it "returns true" do
